Refactor catalog backend integration tests as maven failsafe ITs
