+++ integration-tests/catalog-service-it/pom.xml
@@ -9,14 +9,20 @@
   <artifactId>gs-cloud-catalog-service-it</artifactId>
   <packaging>jar</packaging>
   <description>Catalog client/server integration tests</description>
+  <properties>
+    <surefire.skip>${skipTests}</surefire.skip>
+    <failsafe.skip>${skipTests}</failsafe.skip>
+  </properties>
   <dependencies>
     <dependency>
       <groupId>org.geoserver.cloud</groupId>
       <artifactId>gs-cloud-catalog-client</artifactId>
+      <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.geoserver.cloud</groupId>
       <artifactId>gs-cloud-catalog</artifactId>
+      <scope>test</scope>
     </dependency>
     <dependency>
       <!-- Needed to access GeoServerExtensionsHelper -->
@@ -45,4 +51,25 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <configuration>
+          <skip>${surefire.skip}</skip>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-failsafe-plugin</artifactId>
+        <configuration>
+          <skip>${failsafe.skip}</skip>
+          <forkCount>1</forkCount>
+          <reuseForks>false</reuseForks>
+          <trimStackTrace>false</trimStackTrace>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 </project>
