+++ integration-tests/catalog-service-it/src/test/java/org/geoserver/cloud/integration/catalog/AbstractCatalogBackendIT.java
@@ -1,0 +1,12 @@
+/*
+ * (c) 2020 Open Source Geospatial Foundation - all rights reserved This code is licensed under the
+ * GPL 2.0 license, available at the root application directory.
+ */
+package org.geoserver.cloud.integration.catalog;
+
+import org.geoserver.catalog.CatalogConformanceTest;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+public abstract class AbstractCatalogBackendIT extends CatalogConformanceTest {}
