+++ integration-tests/catalog-service-it/src/test/java/org/geoserver/cloud/integration/catalog/IntegrationTestConfiguration.java
@@ -2,7 +2,7 @@
  * (c) 2020 Open Source Geospatial Foundation - all rights reserved This code is licensed under the
  * GPL 2.0 license, available at the root application directory.
  */
-package org.geoserver.cloud.testconfiguration;
+package org.geoserver.cloud.integration.catalog;
 
 import org.springframework.boot.actuate.autoconfigure.security.servlet.ManagementWebSecurityAutoConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
