+++ integration-tests/catalog-service-it/src/test/java/org/geoserver/cloud/integration/datadirectory/DataDirectoryCatalogIT.java
@@ -5,17 +5,15 @@
 package org.geoserver.cloud.integration.datadirectory;
 
 import org.geoserver.catalog.Catalog;
-import org.geoserver.catalog.CatalogConformanceTest;
 import org.geoserver.catalog.CatalogFacade;
 import org.geoserver.catalog.plugin.CatalogPlugin;
-import org.geoserver.cloud.testconfiguration.IntegrationTestConfiguration;
+import org.geoserver.cloud.integration.catalog.AbstractCatalogBackendIT;
+import org.geoserver.cloud.integration.catalog.IntegrationTestConfiguration;
 import org.geoserver.platform.GeoServerResourceLoader;
 import org.junit.After;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 @SpringBootTest(
     classes = IntegrationTestConfiguration.class,
@@ -25,8 +23,7 @@ import org.springframework.test.context.junit4.SpringRunner;
         "spring.cloud.config.retry.max-attempts=1"
     }
 )
-@RunWith(SpringRunner.class)
-public class DataDirectoryCatalogConformanceTest extends CatalogConformanceTest {
+public class DataDirectoryCatalogIT extends AbstractCatalogBackendIT {
 
     private @Autowired @Qualifier("catalogFacade") CatalogFacade rawCatalogFacade;
     private @Autowired GeoServerResourceLoader resourceLoader;
