+++ pom.xml
@@ -780,6 +780,7 @@
               <version>3.0.0-M5</version>
             </dependency>
           </dependencies>
+          <inherited>true</inherited>
         </plugin>
         <plugin>
           <groupId>org.apache.maven.plugins</groupId>
